backend/requirements.txt
Django>=5.0,<6
psycopg[binary]>=3.1,<4
gunicorn>=22,<24
django-cors-headers>=4.3,<5
Pillow>=10,<12
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
